Extract domain helpers to separate file.
